ajustes

+++ back-end/src/main/java/com/sicredipucrs/AvesDeRapina/dto/BirdDTO.java
@@ -23,7 +23,7 @@ public class BirdDTO {
     private String nameLAT;
 
     @Positive(message = "Size cannot be negative")
-    private Double size;
+    private int size;
     @Size(min = 1, max = 1, message = "Size unit must contain 1 character (M or F)")
     private String gender;
     @NotBlank(message = "Color is mandatory")
